Move http client creation logic to separate service
