+++ AirlineTickets/AirlineTickets.BLL/Interfaces/IHttpClientService.cs
@@ -1,0 +1,7 @@
+﻿namespace AirlineTickets.BLL.Interfaces
+{
+    public interface IHttpClientService
+    {
+        public Task<HttpClient> GetAuthNotificationsClientAsync(CancellationToken cancellationToken);
+    }
+}
